+++ src/lib/helpers/downloadImage.js
